refactor: extract magma trading into separate module
Move all Magma/trading code out of the tapd module into a dedicated
`src/server/modules/api/magma/` module. The tapd module now only handles
local tapd node operations (assets, balances, minting, channels, universe).
