+++ src/client/src/layouts/navigation/Navigation.tsx
@@ -49,12 +49,7 @@ interface NavItem {
   title: string;
   link: string;
   icon: Icon;
-  disabled?: boolean;
-  disabledTooltip?: string;
 }
-
-const TAPD_TOOLTIP =
-  'Enable Trading by running Lightning Terminal with Taproot Assets.';
 
 const mainNav: NavItem[] = [
   { title: 'Home', link: HOME, icon: Home },
@@ -89,40 +84,13 @@ export const Navigation = ({ isBurger, setOpen }: NavigationProps) => {
     capData?.getNodeCapabilities?.capabilities?.includes('taproot_assets') ??
     false;
 
-  const secondaryNav: NavItem[] = secondaryNavItems
-    .filter(item => item.link !== ASSETS || tapdAvailable)
-    .map(item =>
-      item.link === TRADING && !tapdAvailable
-        ? { ...item, disabled: true, disabledTooltip: TAPD_TOOLTIP }
-        : item
-    );
+  const secondaryNav: NavItem[] = secondaryNavItems.filter(
+    item => item.link !== ASSETS || tapdAvailable
+  );
 
   const renderNavButton = (item: NavItem, open = true) => {
     const isActive = nodePath === item.link;
     const NavIcon = item.icon;
-
-    if (item.disabled) {
-      const disabledButton = (
-        <div
-          className={cn(
-            'group relative flex items-center gap-2.5 rounded-md text-xs font-medium cursor-not-allowed opacity-50',
-            open ? 'px-2.5 py-1.5' : 'justify-center p-2'
-          )}
-        >
-          <NavIcon size={15} className="shrink-0" />
-          {open && <span>{item.title}</span>}
-        </div>
-      );
-
-      return (
-        <Tooltip key={item.link}>
-          <TooltipTrigger asChild>{disabledButton}</TooltipTrigger>
-          <TooltipContent side="right" className="text-xs max-w-[200px]">
-            {item.disabledTooltip}
-          </TooltipContent>
-        </Tooltip>
-      );
-    }
 
     const button = (
       <Link key={item.link} to={item.link}>
@@ -158,22 +126,6 @@ export const Navigation = ({ isBurger, setOpen }: NavigationProps) => {
   const renderBurgerNav = (item: NavItem) => {
     const isActive = nodePath === item.link;
     const NavIcon = item.icon;
-
-    if (item.disabled) {
-      return (
-        <Tooltip key={item.link}>
-          <TooltipTrigger asChild>
-            <div className="flex items-center gap-2 px-3 py-2 rounded-md text-sm cursor-not-allowed opacity-50">
-              <NavIcon size={16} />
-              <span>{item.title}</span>
-            </div>
-          </TooltipTrigger>
-          <TooltipContent side="right" className="text-xs max-w-[200px]">
-            {item.disabledTooltip}
-          </TooltipContent>
-        </Tooltip>
-      );
-    }
 
     return (
       <Link key={item.link} to={item.link}>
