+++ src/server/modules/api/magma/magma.gql.ts
