+++ src/server/modules/api/magma/magma.schema.graphql
