+++ src/server/modules/api/tapd/tapd.module.ts
@@ -1,8 +1,5 @@
 import { Module } from '@nestjs/common';
 import { TapdModule } from '../../node/tapd/tapd.module';
-import { FetchModule } from '../../fetch/fetch.module';
-import { NodeModule } from '../../node/node.module';
-import { AmbossModule } from '../amboss/amboss.module';
 import {
   TapdResolver,
   TapAssetGenesisResolver,
@@ -10,7 +7,7 @@ import {
 } from './tapd.resolver';
 
 @Module({
-  imports: [TapdModule, FetchModule, NodeModule, AmbossModule],
+  imports: [TapdModule],
   providers: [TapdResolver, TapAssetGenesisResolver, TapAssetResolver],
 })
 export class TapdApiModule {}
